HH-217505 rename HibernateTestBase to JpaTestBase
